+++ packages/graphquery/src/index.ts
@@ -1,3 +1,4 @@
 export * from './erc20';
 export * from './types';
 export * from './history';
+export * from './allowance';